requirements.txt
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.2
drf_yasg==1.20.0
django-cors-headers==3.2.1
uWSGI==2.0.18
requests==2.26.0
django-json-widget==1.0.1